Add files via upload
Runs a routine to get baselines requiring manual interaction of blocks. Then prints these baselines before attempting to identify the colour of a block using Squared Error. Key parameters are the test cases, delay and number of times each block is scanned (current hardcoded to 5.) Yet to be written is a functionality which bypasses the baseline calibration routine and operates on manually coded expected values for each block's response to a given test set.